Fix S cores mishandling write cycles when rejecting muls
The problem with stalling the way a bus wait cycle forces the execution unit to stall is that it also makes the next execution unit retry the pending write cycle, even if it succeeded. A correct solution, and the one this commit employs, is to synthesise a nop when rejecting a mul exactly the same way the halted core state does.